bug: equalizer not working 20.02.2023

+++ equalizerFeature/src/main/java/com/yes/musicplayer/equalizer/domain/entity/Equalizer.kt
@@ -4,6 +4,6 @@ data class Equalizer (
     val equalizerEnabled:Boolean?=null,
     val currentPreset:Int?=null,
     val presetsNames:List<String>?=null,
-    val bandsLevelRange:ShortArray?=null,
-    val equalizerValues:List<Short>?=null,
+    val bandsLevelRange:IntArray?=null,
+    val equalizerValues:List<Int>?=null,
 )
